+++ code/classeine-lib/clsn/ui/dual_layout_container.cpp
@@ -1,4 +1,4 @@
-#include "dual_container.h"
+#include "dual_layout_container.h"
 
 #include "renderers/dual_container_renderer.h"
 
@@ -6,24 +6,24 @@
 
 namespace clsn::ui
 {
-    dual_container::dual_container()
-    : layout_container<layouts::dual_layout>{"dual_container"}
+    dual_layout_container::dual_layout_container()
+    : layout_container<layouts::dual_layout>{"dual_layout_container"}
     {
     }
 
 
-    auto dual_container::get_orientation() const -> dual_layout_orientation
+    auto dual_layout_container::get_orientation() const -> dual_layout_orientation
     {
         return m_orientation;
     }
 
 
-    void dual_container::set_orientation(dual_layout_orientation orientation)
+    void dual_layout_container::set_orientation(dual_layout_orientation orientation)
     {
         m_orientation = orientation;
     }
 
-    void dual_container::check_if_valid_before_adding(const dual_layout_constraint& constraint) const
+    void dual_layout_container::check_if_valid_before_adding(const dual_layout_constraint& constraint) const
     {
         if (get_count() == 2)
             system::panic("Dual container can only contain two elements");
@@ -33,7 +33,7 @@ namespace clsn::ui
     }
 
 
-    auto dual_container::make_default_renderer() const -> std::unique_ptr<renderer_base>
+    auto dual_layout_container::make_default_renderer() const -> std::unique_ptr<renderer_base>
     {
         return std::make_unique<renderers::dual_container_renderer>();
     }