+++ code/classeine-lib/clsn/ui/renderers/hbox_container_renderer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "clsn/ui/hbox_container.h"
+#include "clsn/ui/hbox_layout_container.h"
 
 #include "clsn/ui/renderers/container_renderer.h"
 
@@ -8,7 +8,7 @@ namespace clsn::ui::renderers
 {
     using namespace clsn::draw;
 
-    class hbox_container_renderer final : public container_renderer<hbox_container>
+    class hbox_container_renderer final : public container_renderer<hbox_layout_container>
     {
     public:
     };