+++ code/classeine-lib/clsn/ui/renderers/xy_container_renderer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "clsn/ui/xy_container.h"
+#include "clsn/ui/xy_layout_container.h"
 
 #include "clsn/ui/renderers/container_renderer.h"
 
@@ -8,7 +8,7 @@ namespace clsn::ui::renderers
 {
     using namespace clsn::draw;
 
-    class xy_container_renderer final : public container_renderer<xy_container>
+    class xy_container_renderer final : public container_renderer<xy_layout_container>
     {
     public:
         void paint(graphics& graphics, const region& region,