+++ code/classeine-lib/clsn/ui/vbox_layout_container.h
@@ -2,16 +2,16 @@
 
 #include "layout_container.h"
 
-#include "layouts/flow_layout.h"
+#include "layouts/vbox_layout.h"
 
 namespace clsn::ui
 {
     using clsn::core::empty;
 
-    class flow_container : public layout_container<layouts::flow_layout>
+    class vbox_layout_container : public layout_container<layouts::vbox_layout>
     {
     public:
-        flow_container();
+        vbox_layout_container();
 
     protected:
         auto make_default_renderer() const -> std::unique_ptr<renderer_base> override;