+++ code/classeine-lib/clsn/ui/xy_layout_container.h
@@ -2,16 +2,16 @@
 
 #include "layout_container.h"
 
-#include "layouts/vbox_layout.h"
+#include "layouts/xy_layout.h"
 
 namespace clsn::ui
 {
     using clsn::core::empty;
 
-    class vbox_container : public layout_container<layouts::vbox_layout>
+    class xy_layout_container : public layout_container<layouts::xy_layout>
     {
     public:
-        vbox_container();
+        xy_layout_container();
 
     protected:
         auto make_default_renderer() const -> std::unique_ptr<renderer_base> override;